// src/types/product.ts
export type Product = {
  id: string;
  uid: string;
  name: string;
  price: number;
  weight: string;
  image: string;
  images: string[];
  category: string;
  categorySlug: string;
  rating: number;
  content: string;
  shelfLife: string;
  isNew: boolean;
  isHit: boolean;
  isPopular: boolean;
  isRecommend: boolean;
};